数据层实现

+++ LCMS/Client/src/main/java/org/businesslogicservice/manageblservice/IncomeTableBLService.java
@@ -2,8 +2,8 @@ package org.businesslogicservice.manageblservice;
 
 public interface IncomeTableBLService {
 	
-	public int getIncome();
+	public long getIncome();
 	
-	public int getPayment ();
+	public long getPayment ();
 
 }
